+++ client/src/main.jsx
@@ -2,6 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import { BrowserRouter } from 'react-router-dom'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
+import { FeatureFlagsProvider } from './contexts/FeatureFlagsContext'
 import { NotificationProvider } from './contexts/NotificationContext'
 import App from './App.jsx'
 import './index.css'
@@ -88,11 +89,13 @@ const queryClient = new QueryClient({
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
     <QueryClientProvider client={queryClient}>
-      <NotificationProvider>
-        <BrowserRouter basename={basename}>
-          <App />
-        </BrowserRouter>
-      </NotificationProvider>
+      <FeatureFlagsProvider>
+        <NotificationProvider>
+          <BrowserRouter basename={basename}>
+            <App />
+          </BrowserRouter>
+        </NotificationProvider>
+      </FeatureFlagsProvider>
     </QueryClientProvider>
   </React.StrictMode>,
 )